% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_locations.R
\name{get_function_locations}
\alias{get_function_locations}
\title{Find where functions are in an R file}
\usage{
get_function_locations(path)
}
\arguments{
\item{path}{The path to an R file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the columns:
\itemize{
\item \code{name}, the name of the function associated with a token.
\item \code{token}, either the name of the function, or an open or closed curly bracket.
\item \code{line}, the line of the code file where the token appears.
}

If the \code{name} and \code{token} is \code{NA} then the row represents an anonymous
function.
}
\description{
Find where functions are in an R file
}
\examples{
\dontrun{

# First specify the path to an R file
file_path <- system.file("test", "test_functions.R", package = "kross")

get_function_locations(file_path)

}

}
